Added support for local publish in doom config
I wrote a function which will look for an export-config.el file in the
current directory. it is currently a very naive function. This function
is bound to "SPC m P l" in my doom emacs configuration. the
export-config.el file must set org-publish-local-alist to its own
configuration alist.

I believe there is a bug in the function currently which causes it to
only work when the org-publish-local-alist is a cons cell rather than an
alist. This will be fixed soon.

The current export options are not comprehensive either, and will
certainly need fixed. I am unsure if I can change it so it uses relative
paths, but that is something that will need to be fixed with the config
as well.