Renamed navbar.org to includes.org
there was a lot of stuff which was repeated fairly often. It is now
consolidated into a few include files.

unfortunately, because I am using relative file paths, the links break
when I import the include file in a subdirectory. My solution is simply
adding a new include file for every subdirectory.