Reorganized CSS and began shaping portfolio.
I made some changes to the overall CSS structure of the webpage. Now,
the color for the content is not stored in the body, but rather in the
div which contains the paragraphs. This allows for different background
colors for different pages.

I plan on using a white background for the portfolio page, which is why
this change needed to take place.