renamed content divs
main-content -> blog-content
link-map -> link-list

I plan to add a couple more styles in the future, but the idea is that
these styles will fully define what the page looks like. Before, all
pages used main-content, but overrode some attributes in a secondary
class. This was annoying, so now each page which has a different format
has a different class.